Define a new memory option, $MEMORY_MAP_EXTENSION.
Glulx has a very simple memory-extension feature, where you declare the
memory map to be larger than the game file. The extra space is filled
out with zeroes when the game starts.

This has been in the spec since the beginning, and interpreters support it
(as far as I know!) but I6 doesn't. This is the I6 patch to support it.