Change the Dictionary directive (currently deprecated) to have the forms:
  Dictionary 'word';
  Dictionary 'word' val1;
  Dictionary 'word' val1 val3;

The first just adds the word to the dictionary, with all flags set to
zero, if it's not already in the dictionary. The second form also sets
the dict_par1 flag to the given value, or bit-or's the value in if the
word already exists. The third form also sets/ors dict_par3.

The values can be numeric literals or constants. They can be 0-255 for
Z-code, 0-65535 for Glulx.