1023: Compiler crashes reporting errors when the source file size reaches 64k lines
Three changes here. First, we replace all the file-line multiplying
factors with a macro, so that it's easy to change. Second, we change
the macro value from 0x10000 to 0x800000.

Finally, we check the range of InputFiles[j-1] in print_preamble() so
that if a bad file-number value creeps in, it doesn't crash the compiler.