First part of work for 1199: Z-spec 1.1 has new header bit
This adds the $ZCODE_HEADER_EXT_WORDS setting, and redefines -Wn to be
a synonym for it. The value defaults to 3, for backwards compatibility
with Inform 6.32 (which always put at least three words in the ext header
whether they were needed or not).

The new behavior is that you can set any value, but if a Unicode translation
table is defined, the value is bumped up to at least three.